Implemented children reparenting for mesh template
We use a mesh template which is a fake node to instance the initial
mesh nodes . This is to ensure the entire tree can be built.

We replace mesh node templates with the real mesh after the
skeleton is available, since this makes it ensure that the fully
built skeleton exists with all bones, all nodes, etc.

If create_mesh was ever moved to the same phase skeleton lookup would be
buggy.

Making skeletons available in a post process means that armature and
skeleton, bone code would need moved to post process step.

Therefore it's simpler to just move create_mesh to run after.