Add unit testing to Godot using DocTest
Implements exit codes into the engine so tests can return their statuses.
Ideally we don't do this, and we use FIXUP logic to 'begin' and 'end' the engine execution for tests specifically.

Since realistically we're initialising the engine here we don't want to do that, since String should not require an engine startup to test a single header. This decouples and lowers the complexity.

Remove legacy unit tests and only use doctest, properly implemented with return code.